requirements.txt
reportlab==4.2.2
svglib==1.5.1
Pillow==10.4.0